+++ src/Core2D/ViewModels/Editor/Tools/ToolEllipse.cs
@@ -23,8 +23,8 @@ namespace Core2D.Editor.Tools
         private Mode _currentMode = Mode.Rectangle;
         private IEllipseShape _ellipse;
         private ToolEllipseSelection _selection;
-        private double _centerX;
-        private double _centerY;
+        private decimal _centerX;
+        private decimal _centerY;
 
         /// <inheritdoc/>
         public string Title => "Ellipse";
@@ -54,13 +54,13 @@ namespace Core2D.Editor.Tools
             throw new NotImplementedException();
         }
 
-        private static void CircleConstrain(IPointShape tl, IPointShape br, double cx, double cy, double px, double py)
+        private static void CircleConstrain(IPointShape tl, IPointShape br, decimal cx, decimal cy, decimal px, decimal py)
         {
-            double r = Max(Abs(cx - px), Abs(cy - py));
-            tl.X = cx - r;
-            tl.Y = cy - r;
-            br.X = cx + r;
-            br.Y = cy + r;
+            decimal r = Max(Abs(cx - px), Abs(cy - py));
+            tl.X = (double)(cx - r);
+            tl.Y = (double)(cy - r);
+            br.X = (double)(cx + r);
+            br.Y = (double)(cy + r);
         }
 
         /// <inheritdoc/>
@@ -68,7 +68,7 @@ namespace Core2D.Editor.Tools
         {
             var factory = _serviceProvider.GetService<IFactory>();
             var editor = _serviceProvider.GetService<IProjectEditor>();
-            (double sx, double sy) = editor.TryToSnap(args);
+            (decimal sx, decimal sy) = editor.TryToSnap(args);
             switch (_currentState)
             {
                 case State.TopLeft:
@@ -83,12 +83,12 @@ namespace Core2D.Editor.Tools
                             editor.Project.CurrentStyleLibrary.Selected :
                             editor.Factory.CreateShapeStyle(ProjectEditorConfiguration.DefaulStyleName);
                         _ellipse = factory.CreateEllipseShape(
-                            sx, sy,
+                            (double)sx, (double)sy,
                             (IShapeStyle)style.Copy(null),
                             editor.Project.Options.DefaultIsStroked,
                             editor.Project.Options.DefaultIsFilled);
 
-                        var result = editor.TryToGetConnectionPoint(sx, sy);
+                        var result = editor.TryToGetConnectionPoint((double)sx, (double)sy);
                         if (result != null)
                         {
                             _ellipse.TopLeft = result;
@@ -112,11 +112,11 @@ namespace Core2D.Editor.Tools
                             }
                             else
                             {
-                                _ellipse.BottomRight.X = sx;
-                                _ellipse.BottomRight.Y = sy;
+                                _ellipse.BottomRight.X = (double)sx;
+                                _ellipse.BottomRight.Y = (double)sy;
                             }
 
-                            var result = editor.TryToGetConnectionPoint(sx, sy);
+                            var result = editor.TryToGetConnectionPoint((double)sx, (double)sy);
                             if (result != null)
                             {
                                 _ellipse.BottomRight = result;
@@ -161,14 +161,14 @@ namespace Core2D.Editor.Tools
         public void Move(InputArgs args)
         {
             var editor = _serviceProvider.GetService<IProjectEditor>();
-            (double sx, double sy) = editor.TryToSnap(args);
+            (decimal sx, decimal sy) = editor.TryToSnap(args);
             switch (_currentState)
             {
                 case State.TopLeft:
                     {
                         if (editor.Project.Options.TryToConnect)
                         {
-                            editor.TryToHoverShape(sx, sy);
+                            editor.TryToHoverShape((double)sx, (double)sy);
                         }
                     }
                     break;
@@ -178,7 +178,7 @@ namespace Core2D.Editor.Tools
                         {
                             if (editor.Project.Options.TryToConnect)
                             {
-                                editor.TryToHoverShape(sx, sy);
+                                editor.TryToHoverShape((double)sx, (double)sy);
                             }
 
                             if (_currentMode == Mode.Circle)
@@ -187,8 +187,8 @@ namespace Core2D.Editor.Tools
                             }
                             else
                             {
-                                _ellipse.BottomRight.X = sx;
-                                _ellipse.BottomRight.Y = sy;
+                                _ellipse.BottomRight.X = (double)sx;
+                                _ellipse.BottomRight.Y = (double)sy;
                             }
                             editor.Project.CurrentContainer.WorkingLayer.InvalidateLayer();
                             Move(_ellipse);
